+++ src/activateMockDebug.ts
@@ -11,7 +11,7 @@ import * as vscode from 'vscode';
 import { WorkspaceFolder, DebugConfiguration, ProviderResult, CancellationToken } from 'vscode';
 import { MockDebugSession } from './mockDebug';
 import { FileAccessor } from './mockRuntime';
-import { getImage } from './tools';
+import { getHtml, getImage } from './tools';
 
 let currentSessionID: string;
 
@@ -179,8 +179,14 @@ export function activateMockDebug(context: vscode.ExtensionContext, factory?: vs
 			if (!targetResource && vscode.window.activeTextEditor) {
 				targetResource = vscode.window.activeTextEditor.document.uri;
 			}
-			//log(currentSessionID);
 			getImage(currentSessionID);
+		}),
+		vscode.commands.registerCommand('extension.mock-debug.getCircuitHtml', (extensionContext: vscode.ExtensionContext, resource: vscode.Uri) => {
+			let targetResource = resource;
+			if (!targetResource && vscode.window.activeTextEditor) {
+				targetResource = vscode.window.activeTextEditor.document.uri;
+			}
+			getHtml(currentSessionID);
 		}),	
 		vscode.commands.registerCommand('extension.mock-debug.toggleFormatting', (variable) => {
 			const ds = vscode.debug.activeDebugSession;